:recycle:refactor: refactor posting detail page 타입 분리하여 위치 이통 사용하지 않는 코드 삭제 Related to: #269
