+++ src/components/PostingDetailPage/types.ts
@@ -1,0 +1,11 @@
+import { User } from '@shared/types/type';
+
+export interface OtherPost {
+  id: string;
+  title: string;
+  content: string;
+  createdAt: string;
+  user: User;
+  isbn: string;
+  likeCount: number;
+}
