+++ src/features/PostDetailPage/types/types.ts
@@ -1,13 +1,35 @@
-import { User } from '@shared/types/type';
-
-export interface OtherPost {
+export interface DbPostingResponse {
   id: string;
-  title: string;
-  content: string;
-  createdAt: string;
-  user: User;
+  created_at: string;
   isbn: string;
-  likeCount: number;
+  user: {
+    id: string;
+    username: string;
+    avatar_url: string;
+  };
+  posting: {
+    title: string;
+    content: string;
+  };
+}
+
+export interface DbBookOtherPostsResponse {
+  id: string;
+  created_at: string;
+  user: {
+    id: string;
+    username: string;
+    avatar_url: string;
+  };
+  isbn: string;
+  posting: {
+    title: string;
+    content: string;
+  };
+  post_like: {
+    user_id: string;
+  }[];
+  like_count: number;
 }
 
 // 댓글 타입
@@ -27,36 +49,10 @@ export interface DbComment {
     avatar_url: string | null;
   };
   likes_count: number;
-  likes: string[]; // 뷰에서 제공하는 좋아요한 사용자 ID 배열
+  likes: string[];
 }
 
 export interface DbCommentCount {
   post_id: string;
   total_comments_count: number;
-}
-
-// 프론트엔드 타입 (카멜 케이스)
-export interface Comment {
-  id: string;
-  createdAt: string;
-  updatedAt: string;
-  postId: string;
-  userId: string;
-  content: string;
-  parentId: string | null;
-  isEdited: boolean;
-  isDeleted: boolean;
-  user: {
-    id: string;
-    username: string;
-    avatarUrl: string | null;
-  };
-  likesCount: number;
-  likes: string[]; // 좋아요한 사용자 ID 배열
-  isLiked: boolean;
-}
-
-export interface CommentCount {
-  postId: string;
-  commentsCount: number;
 }
